+++ MyHeartCounts/Heart Health Dashboard/PastTimedWalkTestResults.swift
@@ -70,14 +70,7 @@ struct PastTimedWalkTestResults: View {
     @ViewBuilder
     private func makeDetailsView(for result: TimedWalkingTestResult) -> some View {
         Form {
-            LabeledContent("Start", value: result.startDate, format: .dateTime)
-            LabeledContent("End", value: result.endDate, format: .dateTime)
-            LabeledContent("Number of Steps", value: result.numberOfSteps, format: .number)
-            LabeledContent(
-                "Distance",
-                value: Measurement<UnitLength>(value: result.distanceCovered, unit: .meters),
-                format: .measurement(width: .abbreviated)
-            )
+            TimedWalkRunTestResultInfo(result: result)
         }
         .navigationTitle(result.test.displayTitle.localizedString())
         .navigationBarTitleDisplayMode(.inline)
