+++ MyHeartCounts/Task Handling/Active Tasks/Timed Walk Test/TimedWalkingTestView.swift
@@ -197,21 +197,7 @@ private struct TimedWalkingTestView: View {
     @ViewBuilder private var resultsSection: some View {
         if !testIsRunning, let result = mostRecentResult {
             Section("Test Complete") {
-                LabeledContent("Date", value: result.startDate, format: .dateTime)
-                LabeledContent("Steps", value: result.numberOfSteps, format: .number)
-                LabeledContent(
-                    "Distance",
-                    value: Measurement<UnitLength>(value: result.distanceCovered, unit: .meters),
-                    format: .measurement(width: .abbreviated)
-                )
-                LabeledContent(
-                    "VO2MAX_ESTIMATE_LABEL",
-                    value: Measurement<UnitOxygenConsumption>(
-                        value: (result.distanceCovered - 504.9) / 44.73,
-                        unit: .mlPerKgPerMin
-                    ),
-                    format: .measurement(width: .abbreviated, usage: .asProvided, numberFormatStyle: .number.precision(.fractionLength(...1)))
-                )
+                TimedWalkRunTestResultInfo(result: result)
             }
         }
     }
