+++ MyHeartCountsUITests/TaskHandlingTests.swift
@@ -6,55 +6,11 @@
 // SPDX-License-Identifier: MIT
 //
 
-import HealthKit
-import SpeziFoundation
 import XCTest
-import XCTestExtensions
 import XCTHealthKit
 
 
 final class TaskHandlingTests: MHCTestCase, Sendable {
-    private var timedWalkTestDistance: String {
-        switch appLocale.measurementSystem {
-        case .us:
-            "2,762 ft"
-        default:
-            "842 m"
-        }
-    }
-    
-    
-    // also tests the Timed Walk Test UI
-    func testCompleteScheduledTaskViaAlwaysAvailableMenu() throws {
-        try launchAppAndEnrollIntoStudy()
-        goToTab(.upcoming)
-        
-        let completionMessage = app.collectionViews.staticTexts["Six-Minute Walk Test, Completed"]
-        
-        do {
-            var numTries = 0
-            while numTries < 10, !app.collectionViews.buttons["Take Test: Six-Minute Walk Test"].waitForExistence(timeout: 2) {
-                app.swipeUp()
-                numTries += 1
-            }
-        }
-        XCTAssert(app.collectionViews.buttons["Take Test: Six-Minute Walk Test"].waitForExistence(timeout: 2))
-        XCTAssertFalse(completionMessage.exists)
-        
-        app.navigationBars["Tasks"].buttons["Perform Always Available Task"].tap()
-        XCTAssert(app.buttons["Six-Minute Walk Test"].waitForExistence(timeout: 2))
-        app.buttons["Six-Minute Walk Test"].tap()
-        XCTAssert(app.buttons["Start Test"].waitForExistence(timeout: 2))
-        app.buttons["Start Test"].tap()
-        handleMotionAndFitnessAccessPrompt(timeout: .seconds(2))
-        XCTAssert(app.staticTexts["Test Complete"].waitForExistence(timeout: 15))
-        XCTAssert(app.staticTexts["Steps, 624"].exists)
-        XCTAssert(app.staticTexts["Distance, \(timedWalkTestDistance)"].exists)
-        app.navigationBars.buttons["Close"].tap()
-        XCTAssert(completionMessage.exists)
-    }
-    
-    
     func testECG() throws {
         try launchAppAndEnrollIntoStudy()
         goToTab(.upcoming)
@@ -79,62 +35,6 @@ final class TaskHandlingTests: MHCTestCase, Sendable {
     }
     
     
-    func testHHDPastTimedWalkTests() throws {
-        try launchAppAndEnrollIntoStudy()
-        goToTab(.heartHealth)
-        app.swipeUp()
-        app.swipeUp()
-        app.swipeUp()
-        app.staticTexts["Past Timed Walk/Run Test Results"].tap()
-        sleep(for: .seconds(1))
-        
-        let testCellPred = NSPredicate(format: "label LIKE 'Six-Minute Walk Test, *'")
-        var numTests: Int {
-            app.buttons.matching(testCellPred).count
-        }
-        let numTestsBefore = numTests
-        
-        if case let button = app.buttons.element(matching: testCellPred),
-           button.waitForExistence(timeout: 2) {
-            button.firstMatch.tap()
-            XCTAssert(app.navigationBars["Six-Minute Walk Test"].waitForExistence(timeout: 2))
-            XCTAssert(app.staticTexts["Number of Steps, 624"].waitForExistence(timeout: 2))
-            XCTAssert(app.staticTexts["Distance, \(timedWalkTestDistance)"].waitForExistence(timeout: 2))
-            app.navigationBars["Six-Minute Walk Test"].buttons["BackButton"].tap()
-        }
-        
-        app.buttons["Six-Minute Walk Test"].tap()
-        
-        app.buttons["Start Test"].tap()
-        handleMotionAndFitnessAccessPrompt(timeout: .seconds(2))
-        XCTAssert(app.staticTexts["Test Complete"].waitForExistence(timeout: 10))
-        XCTAssert(app.staticTexts["Steps, 624"].exists)
-        XCTAssert(app.staticTexts["Distance, \(timedWalkTestDistance)"].exists)
-        app.otherElements["MHC.TimedWalkTestView"].navigationBars.buttons["Close"].tap()
-        let numTestsAfter = numTests
-        XCTAssertEqual(numTestsAfter, numTestsBefore + 1)
-    }
-    
-    
-    func testRecoverTimedWalkTest() throws {
-        try launchAppAndEnrollIntoStudy()
-        goToTab(.upcoming)
-        
-        app.navigationBars["Tasks"].buttons["Perform Always Available Task"].tap()
-        XCTAssert(app.buttons["Six-Minute Walk Test"].waitForExistence(timeout: 2))
-        app.buttons["Six-Minute Walk Test"].tap()
-        XCTAssert(app.buttons["Start Test"].waitForExistence(timeout: 2))
-        app.buttons["Start Test"].tap()
-        handleMotionAndFitnessAccessPrompt(timeout: .seconds(2))
-        XCTAssert(app.staticTexts["Your 6-Minute Walk Test is in progress."].waitForExistence(timeout: 5))
-        app.terminate()
-        try launchAppAndEnrollIntoStudy(
-            skipGoingToHomeTab: true
-        )
-        XCTAssert(app.staticTexts["Your 6-Minute Walk Test is in progress."].waitForExistence(timeout: 5))
-    }
-    
-    
     func testHomeTabTaskSheetLifetime() throws {
         try launchAppAndEnrollIntoStudy()
         goToTab(.home)
@@ -149,20 +49,5 @@ final class TaskHandlingTests: MHCTestCase, Sendable {
         app.activate()
         XCTAssert(app.wait(for: .runningForeground, timeout: 2))
         XCTAssert(dietIntroTextElement.waitForExistence(timeout: 2))
-    }
-}
-
-
-extension MHCTestCase {
-    func handleMotionAndFitnessAccessPrompt(timeout: Duration) {
-        let app = XCUIApplication.springboard
-        let alert = app.alerts.element(matching: "label LIKE %@", "“*” would like to access your Motion & Fitness activity.")
-        if alert.waitForExistence(timeout: timeout.timeInterval) {
-            // wait a little longer. sometimes the "exists" above resolves to true while the alert is still being presented,
-            // in which case the tap is a little too early and doesn't actually get registered.
-            let allowButton = alert.buttons["Allow"]
-            XCTAssert(allowButton.wait(for: \.isHittable, toEqual: true, timeout: 5))
-            allowButton.tap()
-        }
     }
 }
